refactor: rename markdownText
